lab done

+++ labs/K33412/Kondratev_Aleksey/Lab_1/scripts/personal-area.js
@@ -35,11 +35,11 @@ async function getEvents() {
 			myEvents.push(item)
 		}
 	})
-	console.log(myEvents)
 	console.log("myevents", myEvents)
+	console.log("myeventsuser", myUsersEvents)
 	console.log("main", mainId)
 	for (let i = 0; i < myEvents.length; i++) {
-		document.querySelector("#container").innerHTML += getCardHtml(myEvents[i].name, myEvents[i].day, myEvents[i].text, myEvents[i].src, myEvents[i].id, mainId[i].id);
+		document.querySelector("#container").innerHTML += getCardHtml(myEvents[i].name, myEvents[i].day, myEvents[i].text, myEvents[i].src, myEvents[i].id);
 	}
 	// myEvents.forEach(card => {
 	// 	document.querySelector("#container").innerHTML += getCardHtml(card.name, card.day, card.text, card.src, card.id, mainId.id);
@@ -47,18 +47,21 @@ async function getEvents() {
 }
 
 async function deleteCard(event, id) {
-	console.log(event.currentTarget);
-	const link = 'http://localhost:3000/userEvents/' + id;
-	console.log(link);
-	const response = await fetch(link, {
-		method: "DELETE",
-		headers: {
-			'Content-Type': 'application/json'
+	for (let i = 0; i < mainId.length; i++) {
+		if (id === mainId[i].eventId && userId === mainId[i].userId) {
+			const link = 'http://localhost:3000/userEvents/' + mainId[i].id;
+			console.log(link);
+			const response = await fetch(link, {
+				method: "DELETE",
+				headers: {
+					'Content-Type': 'application/json'
+				}
+			})
 		}
-	})
+	}
 }
 
-function getCardHtml(name, day, text, src, id, mainId) {
+function getCardHtml(name, day, text, src, id) {
 	return `
 			<div class="card me-4" data-card-id=${id} style="width: 18rem;">
 				<img src=${src} class="card-img-top" alt="Люди идут в поход" style="height: 11rem;">
@@ -67,7 +70,7 @@ function getCardHtml(name, day, text, src, id, mainId) {
 					<p class="card-text">${text}</p>
 					<p class="card-day">${day}</p>
 					<a href="./event.html" class="btn btn-primary mb-2">На сайт мероприятия</a>
-					<form onsubmit="deleteCard(event, ${mainId})">
+					<form onsubmit="deleteCard(event, ${id})">
 						<button type="submit" class="btn btn-secondary" data-bs-toggle="modal" data-bs-target="#exampleModal">
 							Отписаться
 						</button>
